OffCampusBackEnd/requirements.txt
beautifulsoup4==4.6.0
python-dotenv==0.13.0
requests==2.18.4
Django==3.0.4
google-auth==1.20.0
Pillow==7.2.0
